+++ src/plugin.hpp
@@ -13,3 +13,4 @@ extern Model *modelRotorA;
 extern Model *modelReMoveLite;
 extern Model *modelBolt;
 extern Model *modelInfix;
+extern Model *modelStrip;